requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
fastapi==0.115.12
frozenlist==1.5.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
gspread==6.2.0
h11==0.14.0
httplib2==0.22.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.3.2
murmurhash==1.0.12
numpy==2.2.4
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.2
preshed==3.0.9
price-parser==0.4.0
propcache==0.3.1
pt_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.8.0/pt_core_news_sm-3.8.0-py3-none-any.whl#sha256=c304fa04db3af73cd08a250feacf560506e15a2ec2469bd1b09f06847f6b455c
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.2
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.1
thinc==8.3.6
tqdm==4.67.1
twilio==9.5.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
watchfiles==1.0.4
weasel==0.4.1
websockets==15.0.1
wrapt==1.17.2
yarl==1.18.3
